+++ .github/workflows/codeql-analysis.yml
@@ -17,7 +17,7 @@ on:
       - '**software-full.json'
       - '**software-sources.json'
       - '**update-software.yml'
-      - 'update_software.py'
+      - '**update_software_helper.py'
       - '**software-rules.json'
 #on: [ push, pull_request ]
 jobs:
