+++ tests/html_validator_w3c.py
@@ -4,7 +4,7 @@ import json
 import os
 from pathlib import Path
 import re
-from models import Rating
+from helpers.models import Rating
 from tests.utils import get_friendly_url_name, get_translation,\
     set_cache_file
 from tests.w3c_base import calculate_rating, get_data_for_url,\
